* Slightly reduce code size in bootloader * Improve bootloader code

+++ lib/polyfs.h
@@ -56,7 +56,7 @@ int polyfs_fs_open(polyfs_fs_t *fs);
 int polyfs_check_crc(polyfs_fs_t *fs, void *temp, uint16_t tempsize);
 
 int32_t polyfs_fread(polyfs_fs_t *fs, const struct polyfs_inode *inode,
-	void *ptr, uint32_t offset, uint32_t bytes);
+	void *ptr, uint32_t offset, uint16_t bytes);
 
 int polyfs_opendir(polyfs_fs_t *fs, const struct polyfs_inode *parent,
 	polyfs_readdir_t *rd);
